% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{Bspl.cv}
\alias{Bspl.cv}
\title{Choose number of basis functions in least squares B-splines via crossvalidation}
\usage{
Bspl.cv(Y, X, d.seq, n.folds, plot = FALSE)
}
\arguments{
\item{Y}{a response vector (centered)}

\item{X}{vector of covariate observations}

\item{d.seq}{a sequence of candidate numbers of basis functions}

\item{n.folds}{the number of crossvalidation folds}

\item{plot}{a logical indicating whether to plot crossvalidation output}
}
\value{
the number of basis functions selected by crossvalidation
}
\description{
Choose number of basis functions in least squares B-splines via crossvalidation
}
\examples{
data <- data_gen(n = 200,
                 q = 50,
                 r = .9)

cv.d <- Bspl.cv(Y = data$Y.oracle[,1],
                X = data$X[,1],
                d.seq = 3:12,
                n.folds = 5,
                plot = TRUE)

oracle.Bspl.out <- oracle.Bspl(Y = data$Y.oracle[,1],
                               X = data$X[,1],
                               d = cv.d,
                               plot = TRUE)
}
