% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{oracle.Bspl}
\alias{oracle.Bspl}
\title{Fit simple nonparametric regression model with cubic B-splines}
\usage{
oracle.Bspl(Y, X, d, plot = FALSE, x = NULL, alpha = 0.05)
}
\arguments{
\item{Y}{a response vector (centered)}

\item{X}{vector of covariate observations}

\item{d}{the number functions in the cubic B-spline basis}

\item{x}{a vector of values at which evaluations of the fitted function should be returned/plotted}

\item{a}{logical indicating whether a plot of the fitted function should be generated}
}
\value{
a list containing the fitted function and a vector containing the values of the fitted function at the design points
}
\description{
Fit simple nonparametric regression model with cubic B-splines
}
\examples{
data <- data_gen(n = 200, q = 50, r = .9)

oracle.Bspl.out <- oracle.Bspl(Y = data$Y.oracle[,1],
                               X = data$X[,1],
                               d = 10,
                               plot = TRUE,
                               x = NULL)
}
