% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{pcws.poly}
\alias{pcws.poly}
\title{Helps build the basis functions for Legendre polynomials}
\usage{
pcws.poly(x, left.endpoints, K)
}
\description{
Helps build the basis functions for Legendre polynomials
}
