% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{plot_presmth_Legr}
\alias{plot_presmth_Legr}
\title{plot presmoother when it is made of Legendre polynomials}
\usage{
plot_presmth_Legr(x, true.functions = NULL)
}
\description{
plot presmoother when it is made of Legendre polynomials
}
