% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{resmth.Bspl}
\alias{resmth.Bspl}
\title{Fit simple nonparametric regression model with cubic B-splines}
\usage{
resmth.Bspl(Y, X, d, AAt, sigma.hat, plot = FALSE, x = NULL, alpha = 0.05)
}
\arguments{
\item{Y}{a response vector}

\item{X}{vector of covariate observations}

\item{d}{the number functions in the cubic B-spline basis}
}
\value{
a list containing the fitted function and a vector containing the values of the fitted function at the design points
}
\description{
Fit simple nonparametric regression model with cubic B-splines
}
\examples{
data <- data_gen(n = 200, q = 50, r = .9)

spadd.presmth.Bspl.out <- spadd.presmth.Bspl(X = data$X,
                                             Y = data$Y,
                                             d.pre = 20,
                                             lambda = 1,
                                             eta = 3,
                                             n.foi = 6)

resmth.Bspl.out <- resmth.Bspl(Y = data$Y.oracle[,1],
                               X = data$X[,1],
                               d = 6,
                               AAt = spadd.presmth.Bspl.out$AAt[[1]],
                               sigma.hat = spadd.presmth.Bspl.out$sigma.hat[1],
                               plot = TRUE)
}
