% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{smth.Bspl}
\alias{smth.Bspl}
\title{Fit simple nonparametric regression model with cubic B-splines}
\usage{
smth.Bspl(Y, X, d)
}
\arguments{
\item{Y}{a response vector}

\item{X}{vector of covariate observations}

\item{d}{the number functions in the cubic B-spline basis}
}
\value{
a list containing the fitted function and a vector containing the values of the fitted function at the design points
}
\description{
Fit simple nonparametric regression model with cubic B-splines
}
\examples{
# generate some data
n <- 500
d <- 10
f <- function(x){-sin(x*2)}
X <- runif(n,-2.5,2.5)
noise <- rnorm(n)
Y <- f(X) - mean(f(X)) + noise - mean(noise)

# get nonparametric estimate of f
smth.Bspl.out <- smth.Bspl(Y,X,d)
f.hat <- smth.Bspl.out$f.hat

# plot results
par(mfrow=c(1,1))
plot(Y~X)
plot(f.hat,min(X)+1e-3,max(X)-1e-3,add=TRUE,lwd=1.5,col=rgb(0,0,.545))
x.seq <- seq(min(X),max(X),length=500)
f.cent <- f(x.seq) - mean(f(X))
lines(f.cent~x.seq,lty=2)
}
