% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{spadd.presmth.Bspl.cv}
\alias{spadd.presmth.Bspl.cv}
\title{Choose tuning parameters of desparsified lasso presmoothing estimator with cubic B-splines}
\usage{
spadd.presmth.Bspl.cv(X, Y, d.pre, n.lambda, n.eta, n.folds, plot = FALSE)
}
\arguments{
\item{X}{the design matrix}

\item{Y}{the response vector}

\item{d.pre}{the number of intervals in which to divide the support of each covariate}

\item{n.lambda}{the number of candidate lambda values}

\item{n.eta}{the number of candidate eta values}

\item{n.folds}{the number of crossvalidation folds}
}
\value{
a list with the chosen values of the tuning parameters
}
\description{
Choose tuning parameters of desparsified lasso presmoothing estimator with cubic B-splines
}
\examples{
data <- data_gen(n = 200,q = 50,r = .9)

spadd.presmth.Bspl.cv.out <- spadd.presmth.Bspl.cv(X = data$X,
                                                   Y = data$Y,
                                                   d.pre = 20,
                                                   n.lambda = 25,
                                                   n.eta = 25,
                                                   n.folds = 5,
                                                   plot = TRUE)

spadd.presmth.Bspl.out <- spadd.presmth.Bspl(X = data$X,
                                             Y = data$Y,
                                             d.pre = 20,
                                             lambda = spadd.presmth.Bspl.cv.out$cv.lambda,
                                             eta = spadd.presmth.Bspl.cv.out$cv.eta,
                                             n.foi = 6)

plot_presmth_Bspl(spadd.presmth.Bspl.out)
}
