% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{spadd.presmth.Bspl}
\alias{spadd.presmth.Bspl}
\title{Fit the desparsified lasso presmoothing estimator with cubic B-splines}
\usage{
spadd.presmth.Bspl(X, Y, d.pre, lambda, eta, n.foi)
}
\arguments{
\item{X}{the design matrix}

\item{Y}{the response vector (centered)}

\item{d.pre}{the number of intervals in which to divide the support of each covariate}

\item{lambda}{the tuning parameter for fitting the group lasso estimate for the bias correction}

\item{eta}{the tuning parameter for the group lasso projection of one set of basis functions onto those of the other covariates.}

\item{n.foi}{the number of functions (first columns of \code{X}) for which to compute the desparsified lasso presmoothing estimator.}
}
\value{
a list with the fitted functions etc.
}
\description{
Fit the desparsified lasso presmoothing estimator with cubic B-splines
}
\examples{
data <- data_gen(n = 200, q = 10, r = .5)

spadd.presmth.Bspl.out <- spadd.presmth.Bspl(X = data$X,
                                             Y = data$Y,
                                             d.pre = 20,
                                             lambda = 1,
                                             eta = 3,
                                             n.foi = 6)

plot_presmth_Bspl(spadd.presmth.Bspl.out,
                  true.functions = list( f = data$f,
                                         X = data$X))
}
