% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaddinf.R
\name{spadd.presmth.Legr}
\alias{spadd.presmth.Legr}
\title{Fit the desparsified lasso presmoothing estimator with Legendre polynomials}
\usage{
spadd.presmth.Legr(X, Y, d.pre, lambda, eta, n.foi, K = 1)
}
\arguments{
\item{X}{the design matrix}

\item{Y}{the response vector}

\item{d.pre}{the number of intervals in which to divide the support of each covariate}

\item{lambda}{the tuning parameter for fitting the group lasso estimate for the bias correction}

\item{eta}{the tuning parameter for the group lasso projection of one set of basis functions onto those of the other covariates.}

\item{n.foi}{the number of functions (first columns of \code{X}) for which to compute the desparsified lasso presmoothing estimator.}

\item{K}{the order of the Legendre polynomials. E.g. \code{K=0} fits piecwise constant, \code{K=1} fits piecewise linear functions.}
}
\value{
a list with the fitted functions etc.
}
\description{
Fit the desparsified lasso presmoothing estimator with Legendre polynomials
}
\examples{
data <- data_gen(n = 100,q = 10,r = .5)

spadd.presmth.Legr.out <- spadd.presmth.Legr(X = data$X,
                                             Y = data$Y,
                                             d.pre = 10,
                                             lambda = .1,
                                             eta = 2,
                                             n.foi = 6,
                                             K = 1)

plot_presmth_Legr(x = spadd.presmth.Legr.out,
                  true.functions = list( f = data$f, X = data$X))
}
